Load SSNAP data from CSV file into memory

This function imports data from a csv file exported by www.strokeaudit.org.

ssnapcsv(filename, version, ignore_fake_team_data = TRUE,
  dont_warn_if_cols_missing = FALSE, dates_munged_by_excel = FALSE)

Arguments

filename
Filename of the CSV file to load.

version
A version string for the CSV file (currently we only advise using "3.1.6", but any string is accepted without error at present.)

ignore_fake_team_data
A boolean. By default (TRUE) any team codes known to be fake or test team codes are skipped. If FALSE, all teams are exported.

dont_warn_if_cols_missing
suppress warnings if columns are missing. For SSNAP data most imports will not contain all columns and you will be warned which ones are missing if this is FALSE. When TRUE those warnings will be suppressed - at the possible expense of missing genuine errors. So call this function with dont_warn_if_cols_missing set to TRUE if you are debugging any data import problems.

dates_munged_by_excel
If you open a CSV file in Excel and save it (eg. to make a simple edit or remove rows) Excel decides to change the date format into something else. By default we use the standardised date format used by the SSNAP server. If however you get errors because Excel has managed to munge the file, then set dates_munged_by_excel to TRUE.

Value

A tibble containing the imported SSNAP data.

Details

A version string for the SSNAP file is mandated. This is because we don't have that information 'in the file'. At present that version is 3.1.6. Keeping track of the CSV versions allows us to convert obsolete versions of the data to be safe to analyse. This is likely to be deprecated if/when we can be selective over which columns we are loading.

Most SSNAP data exporting all teams will include a number of 'test' or 'fake' teams. By default, we ignore these teams. If however you need to include these teams (for example to test some new code,) you can set ignore_fake_team_data to FALSE.

There are various versions of the SSNAP CSV file, some versions have renamed columns which we need to handle by loading 'both' possible column names. On the current version of SSNAPInterface, this generates warnings when loading. You can suppress these warnings by setting dont_warn_if_cols_missing to TRUE. In future, it is hoped further development of SSNAPInterface will render this need obsolete.

The normal CSV file follows the industry standard. Excel does not follow these standards, and so if you edit the CSV file in Excel it does not usually load. dates_munged_by_excel allows you to load a file which Excel has 'munged'.

The data from the CSV file is optimised as described in the overview.

S1PatientClockStartDateTime, S1TeamClockStartDateTime, S7DeathDate and S7TeamClockStopDateTime are added so that we can select cohorts using these fields.

Future development roadmap

In future we also intend to pass the CSV file a list of the columns that we want using SSNAPStats logic which will select the columns based upon the measures. This should make processing times much quicker, and will deprecate the version string.

We don't yet handle zip file encryption - but it should be possible to pass in encrypted zips of CSV at a later stage, improving data security by allowing the data to be stored in an encrypted format.

In the more distant future dropping CSV altogether in favour of encouraging direct access to the database would be preferable as it would be significantly more secure and reduce the risk of many copies of SSNAP data existing on users' machines.