Bit-field flags for 'No But' reasons for not thrombolysing a patient

In the raw data these are represented by a non-exclusive series of Booleans. To save space and improve processing time, rather than store the tPA No Buts as a series of TRUE/FALSE values, we can store it as a binary number - where 0 is FALSE and 1 is TRUE. Each parameter is therefore given a value of 1^X, where X is incremented from 0.

tPANoBut

Format

An object of class numeric of length 10.

Value

a named numeric vector of thrombolysis reasons and their representative value within the bitfield.

Details

tPANoBut is a vector of named integer values describing which binary bit in a number represents which thrombolysis reason so rather than having numeric values scattered in code which may be unclear or need commenting every time you can refer to a named vector item.

This allows us to take a few mathematical shortcuts.

To find if a value is TRUE you want to do a logical AND against the value.

To find everyone who had no 'no buts' you just look for 0; everyone who had at least one no but has a value > 0.